break out mapping into one file per interface
